merge

+++ xanmod-install-headers.sh
@@ -10,7 +10,7 @@ tyblue()                           #天依蓝
 {
     echo -e "\\033[36;1m${*}\\033[0m"
 }
-green()                            #水鸭青
+green()                            #原谅绿
 {
     echo -e "\\033[32;1m${*}\\033[0m"
 }
@@ -22,11 +22,15 @@ red()                              #姨妈红
 {
     echo -e "\\033[31;1m${*}\\033[0m"
 }
+blue()                             #蓝色
+{
+    echo -e "\\033[34;1m${*}\\033[0m"
+}
 #检查基本命令
 check_base_command()
 {
     local i
-    local temp_command_list=('bash' 'true' 'false' 'exit' 'echo' 'test' 'free' 'sort' 'sed' 'awk' 'grep' 'cut' 'cd' 'rm' 'cp' 'mv' 'head' 'tail' 'uname' 'tr' 'md5sum' 'tar' 'cat' 'find' 'type' 'command' 'kill' 'pkill' 'wc' 'ls')
+    local temp_command_list=('bash' 'true' 'false' 'exit' 'echo' 'test' 'sort' 'sed' 'awk' 'grep' 'cut' 'cd' 'rm' 'cp' 'mv' 'head' 'tail' 'uname' 'tr' 'md5sum' 'cat' 'find' 'type' 'command' 'wc' 'ls' 'mktemp' 'swapon' 'swapoff' 'mkswap' 'chmod' 'chown' 'export')
     for i in ${!temp_command_list[@]}
     do
         if ! command -V "${temp_command_list[$i]}" > /dev/null; then
@@ -35,6 +39,36 @@ check_base_command()
             exit 1
         fi
     done
+}
+#安装单个重要依赖
+test_important_dependence_installed()
+{
+    local temp_exit_code=1
+    if LANG="en_US.UTF-8" LANGUAGE="en_US:en" dpkg -s "$1" 2>/dev/null | grep -qi 'status[ '$'\t]*:[ '$'\t]*install[ '$'\t]*ok[ '$'\t]*installed[ '$'\t]*$'; then
+        if LANG="en_US.UTF-8" LANGUAGE="en_US:en" apt-mark manual "$1" | grep -qi 'set[ '$'\t]*to[ '$'\t]*manually[ '$'\t]*installed'; then
+            temp_exit_code=0
+        else
+            red "安装依赖 \"$1\" 出错！"
+            green  "欢迎进行Bug report(https://github.com/kirin10000/Xray-script/issues)，感谢您的支持"
+            yellow "按回车键继续或者Ctrl+c退出"
+            read -s
+        fi
+    elif apt -y --no-install-recommends install "$1"; then
+        temp_exit_code=0
+    else
+        apt update
+        apt -y -f install
+        apt -y --no-install-recommends install "$1" && temp_exit_code=0
+    fi
+    return $temp_exit_code
+}
+check_important_dependence_installed()
+{
+    if ! test_important_dependence_installed "$@"; then
+        red "重要组件\"$1\"安装失败！！"
+        yellow "按回车键继续或者Ctrl+c退出"
+        read -s
+    fi
 }
 ask_if()
 {
@@ -47,8 +81,12 @@ ask_if()
     [ $choice == y ] && return 0
     return 1
 }
-check_base_command
 
+if [[ -d "/proc/vz" ]]; then
+    red "Error: Your VPS is based on OpenVZ, which is not supported."
+    exit 1
+fi
+check_base_command
 if [[ "$(type -P apt)" ]]; then
     if [[ "$(type -P dnf)" ]] || [[ "$(type -P yum)" ]]; then
         red "同时存在apt和yum/dnf"
@@ -59,13 +97,10 @@ else
     red "xanmod内核仅支持Debian系的系统，如Ubuntu,Debian,deepin,UOS"
     exit 1
 fi
-
 if [ "$EUID" != "0" ]; then
     red "请用root用户运行此脚本！！"
     exit 1
 fi
-
-[[ -d "/proc/vz" ]] && red "Error: Your VPS is based on OpenVZ, which is not supported." && exit 1
 
 menu()
 {
@@ -99,31 +134,18 @@ check_mem()
     fi
 }
 
-check_important_dependence_installed()
-{
-    if dpkg -s "$1" > /dev/null 2>&1; then
-        apt-mark manual "$1"
-    elif ! apt -y --no-install-recommends install "$1"; then
-        apt update
-        if ! apt -y --no-install-recommends install "$1"; then
-            red "重要组件\"$1\"安装失败！！"
-            yellow "按回车键继续或者Ctrl+c退出"
-            read -s
-        fi
-    fi
-}
-
 remove_other_kernel()
 {
-    yellow "卸载过程中如果弹出对话框，请选择NO！"
-    yellow "卸载过程中如果弹出对话框，请选择NO！"
-    yellow "卸载过程中如果弹出对话框，请选择NO！"
-    tyblue "按回车键以继续。。"
-    read -s
+    local temp_file
+    temp_file="$(mktemp)"
+    dpkg --list > "$temp_file"
+    local kernel_list_headers
+    kernel_list_headers=($(awk '{print $2}' "$temp_file" | grep '^linux-headers'))
     local kernel_list_image
-    kernel_list_image=($(dpkg --list | awk '{print $2}' | grep '^linux-image'))
+    kernel_list_image=($(awk '{print $2}' "$temp_file" | grep '^linux-image'))
     local kernel_list_modules
-    kernel_list_modules=($(dpkg --list | awk '{print $2}' | grep '^linux-modules'))
+    kernel_list_modules=($(awk '{print $2}' "$temp_file" | grep '^linux-modules'))
+    rm "$temp_file"
     local i
     local ok_install=0
     for ((i=${#kernel_list_image[@]}-1;i>=0;i--))
@@ -151,10 +173,7 @@ remove_other_kernel()
             return 1
         fi
     fi
-    local exit_code=1
     if [ $install_headers -eq 1 ]; then
-        local kernel_list_headers
-        kernel_list_headers=($(dpkg --list | awk '{print $2}' | grep '^linux-headers'))
         ok_install=0
         for ((i=${#kernel_list_headers[@]}-1;i>=0;i--))
         do
@@ -167,21 +186,43 @@ remove_other_kernel()
             red "内核可能安装失败！不卸载"
             return 1
         fi
+    fi
+    if [ ${#kernel_list_image[@]} -eq 0 ] && [ ${#kernel_list_modules[@]} -eq 0 ] && ([ $install_headers -eq 0 ] || [ ${#kernel_list_headers[@]} -eq 0 ]); then
+        red "未发现可卸载内核！不卸载"
+        return 1
+    fi
+    yellow "卸载过程中如果弹出对话框，请选择NO！"
+    yellow "卸载过程中如果弹出对话框，请选择NO！"
+    yellow "卸载过程中如果弹出对话框，请选择NO！"
+    tyblue "按回车键以继续。。"
+    read -s
+    local exit_code=1
+    if [ $install_headers -eq 1 ]; then
         apt -y purge "${kernel_list_image[@]}" "${kernel_list_modules[@]}" "${kernel_list_headers[@]}" && exit_code=0
     else
         apt -y purge "${kernel_list_image[@]}" "${kernel_list_modules[@]}" && exit_code=0
     fi
-    [ $exit_code -ne 0 ] && red "卸载失败！" && exit 1
-    apt-mark manual "^grub"
+    if [ $exit_code -eq 0 ]; then
+        apt-mark manual "^grub"
+        green "卸载完成"
+    else
+        apt -y -f install
+        apt-mark manual "^grub"
+        red "卸载失败！"
+    fi
 }
 
 main()
 {
     menu
+    check_important_dependence_installed procps
     check_mem
     check_important_dependence_installed gnupg1
     check_important_dependence_installed wget
     check_important_dependence_installed ca-certificates
+    check_important_dependence_installed initramfs-tools
+    local temp_xanmod_apt_source=0
+    [[ -f '/etc/apt/sources.list.d/xanmod-kernel.list' ]] && temp_xanmod_apt_source=1
     echo 'deb http://deb.xanmod.org releases main' | tee /etc/apt/sources.list.d/xanmod-kernel.list
     if ! wget -qO - https://dl.xanmod.org/gpg.key | apt-key --keyring /etc/apt/trusted.gpg.d/xanmod-kernel.gpg add -; then
         red "添加源失败！"
@@ -192,7 +233,7 @@ main()
         exit 1
     fi
     local temp_list
-    temp_list=($(apt-cache depends "$install" | grep -i "Depends:" | awk '{print $2}'))
+    temp_list=($(LANG="en_US.UTF-8" LANGUAGE="en_US:en" apt-cache depends "$install" | grep -i "Depends:" | awk '{print $2}'))
     local i
     for i in ${!temp_list[@]}
     do
@@ -214,9 +255,17 @@ main()
     else
         apt -y --no-install-recommends install "${install_image_list[@]}" "${install_modules_list[@]}" "${install_headers_list[@]}" && exit_code=0
     fi
-    [ $exit_code -ne 0 ] && red "安装失败！" && exit 1
-    ask_if "是否卸载其它内核？(y/n)" && remove_other_kernel
+    [ $exit_code -ne 0 ] && apt -y -f install
+    if [ $temp_xanmod_apt_source -eq 0 ]; then
+        rm /etc/apt/sources.list.d/xanmod-kernel.list
+        apt update
+    fi
+    if [ $exit_code -ne 0 ]; then
+        red "安装失败！"
+        exit 1
+    fi
     green "安装完成"
+    ask_if "是否卸载其它内核？(y/n)" && remove_other_kernel
     yellow "系统需要重启"
     if ask_if "现在重启系统? (y/n)"; then
         reboot
